requirements.txt
alembic @ file:///home/conda/feedstock_root/build_artifacts/alembic_1730767847981/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
Bottleneck @ file:///croot/bottleneck_1707864210935/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
colorlog @ file:///home/conda/feedstock_root/build_artifacts/colorlog_1730249771744/work
comm @ file:///croot/comm_1709322850197/work
contourpy @ file:///opt/conda/conda-bld/contourpy_1663827406301/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
et-xmlfile==1.1.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fonttools @ file:///croot/fonttools_1713551344105/work
greenlet @ file:///home/conda/feedstock_root/build_artifacts/greenlet_1648882382141/work
idna @ file:///croot/idna_1714398848350/work
imbalanced-learn @ file:///home/conda/feedstock_root/build_artifacts/imbalanced-learn_1728063863984/work
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1725921340658/work
ipykernel @ file:///croot/ipykernel_1728665589812/work
ipython @ file:///croot/ipython_1691532092695/work
jedi @ file:///croot/jedi_1721058342488/work
joblib @ file:///croot/joblib_1718217211762/work
jupyter_client @ file:///croot/jupyter_client_1699455897726/work
jupyter_core @ file:///croot/jupyter_core_1718818295206/work
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
Mako @ file:///home/conda/feedstock_root/build_artifacts/mako_1715711344987/work
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1648737563195/work
matplotlib @ file:///croot/matplotlib-suite_1693812469450/work
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
numexpr @ file:///croot/numexpr_1683221822650/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
openpyxl @ file:///croot/openpyxl_1721752957391/work
optuna @ file:///home/conda/feedstock_root/build_artifacts/optuna_1744686529628/work
packaging @ file:///croot/packaging_1720101850331/work
pandas==1.5.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.3.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
pooch @ file:///croot/pooch_1695850093751/work
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
Pygments @ file:///croot/pygments_1684279966437/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1648757091578/work
pyzmq @ file:///croot/pyzmq_1705605076900/work
requests @ file:///croot/requests_1721410876868/work
scikit-learn @ file:///croot/scikit-learn_1694788527225/work
scipy==1.10.1
seaborn @ file:///croot/seaborn_1718302919398/work
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
SQLAlchemy @ file:///home/conda/feedstock_root/build_artifacts/sqlalchemy_1651017965248/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
threadpoolctl @ file:///croot/threadpoolctl_1719407800858/work
torch==1.9.0
torchaudio==0.9.0a0+33b2469
torchvision==0.10.0
tornado @ file:///croot/tornado_1718740109488/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1732497199771/work
traitlets @ file:///croot/traitlets_1718227057033/work
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
urllib3 @ file:///croot/urllib3_1727769808118/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
zipp @ file:///croot/zipp_1729012354496/work
